refactor(rag): consolidate RAG MCP tooling under app/rag/mcp
Move the RAG MCP package (app/rag_mcp) and the LangChain wrapper layer
(app/rag/agent_tools.py) into a single app/rag/mcp/ subpackage, so the RAG
module is self-contained and the layout matches how the code is actually used.
Tests move to the mirrored tests/rag/mcp/. Pure move + import refactor, no
behaviour change.
