+++ app/core_plugins/chat/tests/test_intent_router.py
@@ -36,7 +36,7 @@ class TestRAGIntentRouterDecide:
         llm = _make_llm(decision)
         router = RAGIntentRouter(llm=llm)
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             result = await router.decide(
@@ -55,7 +55,7 @@ class TestRAGIntentRouterDecide:
         llm = _make_llm(decision)
         router = RAGIntentRouter(llm=llm)
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             result = await router.decide(
@@ -77,7 +77,7 @@ class TestRAGIntentRouterDecide:
 
         router = RAGIntentRouter(llm=llm)
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             with pytest.raises(RAGIntentRouterError):
@@ -107,7 +107,7 @@ class TestRAGIntentRouterDecide:
 
         router = RAGIntentRouter(llm=llm)
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             with pytest.raises(RAGIntentRouterError):
@@ -125,7 +125,7 @@ class TestRAGIntentRouterDecide:
         router = RAGIntentRouter(llm=llm)
         query_text = "summarize chapter 3"
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             await router.decide(
@@ -153,7 +153,7 @@ class TestRAGIntentRouterDecide:
         router = RAGIntentRouter(llm=llm)
         file_name = "textbook.pdf"
 
-        with patch("app.rag_mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
+        with patch("app.rag.mcp.tools.get_document_structure", new_callable=AsyncMock) as mock_struct:
             mock_struct.return_value = []
 
             await router.decide(
