+++ app/rag/mcp/tools.py
@@ -8,9 +8,9 @@ from app.lib.db import session_scope
 from app.lib.log import get_logger
 from app.models.drive import DriveFile
 from app.rag.db_models import DocumentChunk
+from app.rag.mcp.schemas import ChunkStats, DocumentSection, FileInfo, FileMetadata, SectionKey
 from app.rag.types import RagStatus
 from app.rag.utils import resolve_source_name
-from app.rag_mcp.schemas import ChunkStats, DocumentSection, FileInfo, FileMetadata, SectionKey
 
 logger = get_logger(__name__)
